# specs/001-haruhanjang-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 하루한장 API
  version: 0.1.0
  description: |
    사장님 PWA(S1~S8)와 기관 대시보드(S7)가 쓰는 REST API.
    근거: spec.md FR-xxx, design/SD_01 IPO, SD_02 화면, SD_03 테이블·뷰.

    공통 규칙
    - 모든 경로는 `/api` 아래. 인증은 HttpOnly 쿠키(`hrh_session`, 서명 JWT).
    - 날짜는 KST 기준 `YYYY-MM-DD`.
    - 게이트 차단은 HTTP 409 + `GateBlocked` 본문. 화면은 `gate` 코드로 SD_02 §11-2 표준 문구를 고른다.
    - 사장님 API는 로그인한 계정의 가게 데이터만 다룬다(FR-091). 경로에 store_id를 받지 않는다.
    - 기관 API(`/api/org/*`)는 익명 집계 뷰만 조회한다(FR-076, research R7).
    - 인가(RBAC): 각 경로의 권한은 data-model §7-2. 권한이 없으면 403 Forbidden, 게이트 차단은 409.
servers:
  - url: /api
...
paths:
...
  /admin/thresholds:
    get:
      tags: [admin]
      summary: 기준값 목록 (권한 threshold.manage)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Threshold' } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/thresholds/{key}:
    put:
      tags: [admin]
      summary: 기준값 설정·비우기 (권한 threshold.manage). null 이면 해당 게이트 차단
      parameters:
        - { in: path, name: key, required: true, schema: { type: string, enum: [g3_min_records, g5_min_stores, alert_window_days, alert_min_decline] } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [value], properties: { value: { type: number, nullable: true, minimum: 0 } } }
      responses:
        '204': { description: '저장' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/codes/{codeKind}:
    parameters:
      - { in: path, name: codeKind, required: true, schema: { type: string, enum: [regions, business-types, sales-bands] } }
    get:
      tags: [admin]
      summary: 코드 목록 (권한 code.read)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin]
      summary: 코드 추가·수정 (권한 code.manage). 사용 중인 코드 삭제는 제공하지 않음
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '204': { description: '저장' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/orgs:
    get:
      tags: [admin]
      summary: 기관 목록·계약 상태 (권한 org.manage)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/OrgSummary' } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin]
      summary: 기관 등록 (권한 org.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [orgName, jurisdiction], properties: { orgName: { type: string }, jurisdiction: { type: array, items: { type: string } } } }
      responses:
        '201': { description: '등록' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/orgs/{orgId}/contract:
    patch:
      tags: [admin]
      summary: 계약 상태·만료일 변경 (권한 contract.manage) — G8 해제 경로
      parameters:
        - { in: path, name: orgId, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [status], properties: { status: { type: string, enum: [active, expired] }, endDate: { type: string, format: date, nullable: true } } }
      responses:
        '204': { description: '저장' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/orgs/{orgId}/jurisdiction:
    put:
      tags: [admin]
      summary: 관할 지역 교체 (권한 org.manage)
      parameters:
        - { in: path, name: orgId, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [regionCodes], properties: { regionCodes: { type: array, items: { type: string } } } }
      responses:
        '204': { description: '저장' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/orgs/{orgId}/accounts:
    post:
      tags: [admin]
      summary: 기관 계정 발급 (권한 org.manage) — 임시 비밀번호는 응답으로 한 번만 반환
      parameters:
        - { in: path, name: orgId, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [loginId], properties: { loginId: { type: string } } }
      responses:
        '201':
          description: '발급, org_viewer 역할 자동 부여'
          content:
            application/json:
              schema: { type: object, properties: { orgAccountId: { type: integer }, temporaryPassword: { type: string } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/audit/{auditKind}:
    get:
      tags: [admin]
      summary: 감사 이력 (권한 audit.read). 개별 사장님 기록은 포함하지 않음
      parameters:
        - { in: path, name: auditKind, required: true, schema: { type: string, enum: [gate-events, exports, org-queries, anon-cells, rbac-grants] } }
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/staff:
    get:
      tags: [admin]
      summary: 운영 인력 목록·역할 (권한 role.manage)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/StaffSummary' } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [admin]
      summary: 운영 인력 계정 생성 (권한 role.manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [loginId, displayName, roles], properties: { loginId: { type: string }, displayName: { type: string }, roles: { type: array, items: { type: string, enum: [data_manager, operator, auditor, admin] } } } }
      responses:
        '201':
          description: '생성, 임시 비밀번호 1회 반환'
          content:
            application/json:
              schema: { type: object, properties: { staffAccountId: { type: integer }, temporaryPassword: { type: string } } }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/staff/{staffId}:
    patch:
      tags: [admin]
      summary: 역할 교체·비활성화 (권한 role.manage). 마지막 admin 회수·자기 admin 회수는 409
      parameters:
        - { in: path, name: staffId, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, properties: { roles: { type: array, items: { type: string, enum: [data_manager, operator, auditor, admin] } }, isActive: { type: boolean } } }
      responses:
        '204': { description: '저장, rbac_grant_event 기록' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { description: '마지막 admin 회수 또는 자기 admin 회수' }
components:
  responses:
...
    Forbidden:
      description: 권한 없음 (RBAC). 게이트 차단(409)과 구분한다
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: forbidden }
              permission: { type: string, example: threshold.manage }
...
  schemas:
...
    Threshold:
      type: object
      properties:
        key: { type: string }
        value: { type: number, nullable: true }
        unitLabel: { type: string }
        updatedAt: { type: string, format: date-time }
    OrgSummary:
      type: object
      properties:
        orgId: { type: integer }
        orgName: { type: string }
        contractStatus: { type: string, enum: [active, expired] }
        contractEndDate: { type: string, format: date, nullable: true }
        jurisdiction: { type: array, items: { type: string } }
    StaffSummary:
      type: object
      properties:
        staffAccountId: { type: integer }
        loginId: { type: string }
        displayName: { type: string }
        isActive: { type: boolean }
        roles: { type: array, items: { type: string } }
...